refactor read_medpc into separate file
